Restart SNTP client when we have an IPv4 or IPv6 address
This fixes the restarting of the SNTP client when the network
interface has a valid IP address. This was broken in changeset
9246ef for IPv4.

Note that for IPv6 the SNTP server can take up to 15 seconds to
be used due to the async nature of stateless DHCPv6 and SLAAC.
For IPv4 it takes effect immediately as the address and SNTP
server come in the same packet.

Tests:
 - Verify operation on dual-stack IPv4/IPv6:
   - DHCPv6 only network (no DHCPv4 server present)
   - DHCPv4 and DHCPv6 network
 - Verify operation on IPv4 only stack
 - Check SNTP restart with LwIP logging enabled
 - Verify time is requested and received via Wireshark
 - Verify system time is set

Signed-off-by: David J. Fiddes <D.J@fiddes.net>