Allow overriding netif->input (#64)
Instead of statically calling ethernet_input() (which is actually
defined to be ethernet_input_LWIP2), store a pointer to it in
netif->input and call that.

This makes it easy to hook into the processing of input packets and
implement a little firewall. Of course any code doing this should save
the original pointer and call it when applicable. For an example, see:

  https://github.com/martin-ger/esp_wifi_repeater/blob/2499913094a707b79ea4e357273a32755e31c723/user/user_main.c#L693-L697

Also remove various lines where netif->input would be reset. Set it
once during netif_add(), and then trust that it wouldn't get changed
without reason.